refactor: #25 - 폴더명 변경
